chore(genie): hopefully the last big refactor
Wasn't happy with the organization which started getting messy and
lacked clarity as components started to get added.  I've moved the
configuration details into the individual packages in order to create
the various objects during the configuration stage.  There's still some
cleanup left to do in terms of initializing some of the objects, but the
bulk of the changes have been made and are functional.