feat(run-once): add run-once option to generator
Adds an option to only run a single time.  By default logging is
enabled as it is assumed that the option will be used locally while
developing resources/events.  In working through this, I discovered some
sync issues on the channel shutdown, so I changed up the pattern a bit
so we are using the closing the send channel instead of using a stop
channel.