chore(sinks): remove sinks from the event config
This removes the sink reference from the event config in favor of
specifying the sink on the command line, where I expect it will be used
more often since this is a test/dev tool. As part of the change to
simplify things, I also pulled the sink dependency from the manager and
minimized the resources that it needed to know about.  The generator
command will now get the sink and pass the channel to the manager on
startup.  Manager is now a simple wrapper around event startup and only
needs to track the events that it has started.

There's a possiblity that I'll follow up this PR with a new one that
allows the user to specify multiple sinks to start which would also
required sending multiple channels to the events, but for now this works
sufficiently well.  In the case that I do that, I'll move the sink
initialization and start into the sink package and out of the generator.