feat(integer_range): add distribution types
Adds the distribution settings to the integer range resource.  By
default the range still returns random numbers with a uniform
distribution, but now the option exists to return integers in a normal
distribution.  The mean and the size of the standard deviation can be
set in order to manipulate range.  While I was in there I added the
ability to limit to a specific event on the command line, it was
something that I had wanted to add, but just never did it.