feat(float_range): add float_range resource
Adds a new float_range resource.  I may revisit this later and merge
both the integer and float ranges into a number range.  For now it made
sense to split it off since there were significant differences between
the way the two are managed.  That being said, a number range that is
based off of the float resource and just casts to an integer would add
additional features. Like the integer range, float ranges
support the uniform and normal distributions, but also support an
exponential distribution.  Formatting is also supported.

I've made a note on the integer range to remove padding from the
resource and create a filter for it.  Once that's done, I'll look more
closely at merging all the number resources.